refactor(workflows): collapse follow-up modal into function component
All modal events already bubbled to WorkflowRunnerLive — the live_component
added an indirection hop (phx-target + send/handle_info) without any local
state that justified a process of its own. Moving :selected_type up to the
parent lets the modal render as a stateless function component and the
parent handle the four events directly.
